requirements.txt
cffi               == 1.15.1
click              == 8.1.3
colorama           == 0.4.6
cryptography       == 38.0.4
Flask              == 2.2.2
Flask-Cors         == 3.0.10
Flask-JWT-Extended == 4.4.4
Flask-MySQL        == 1.5.2
pycparser          == 2.21
PyJWT              == 2.6.0
PyMySQL            == 1.0.2
six                == 1.16.0
Werkzeug           == 2.2.2
zipp               == 3.11.0
requests           == 2.28.1
urllib3            == 1.26.13